move utility method from AbstractPluginUITest to UIUtils
